feat: add first flag annotation to instancers
